requirements.txt
# Core
Django>=4.2,<5.1
asgiref>=3.7,<4
sqlparse>=0.4,<0.6
tzdata>=2024.1
pytz>=2024.1

# Environment
django-environ==0.12.0
python-decouple==3.8

# Django utils
django-extensions==4.1
django-cors-headers==4.9.0
django-filter>=23.1,<25
django-debug-toolbar>=4.3,<5
whitenoise>=6.6,<7
dj-database-url>=2.1,<3

# REST
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.28.0
drf-yasg

# Auth / JWT
PyJWT>=2.8,<3

# Media / Images
Pillow>=10.0,<12
imagekitio==4.1.0
pydantic<2

# DB
psycopg2-binary==2.9.11

# Storage
django-storages[boto3]==1.14.4

# Utils
requests>=2.31,<3
PyYAML>=6.0,<7
certifi

# Server
gunicorn>=21,<22
